fix compiler warning
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/branches/v01_00_xx@4493 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ ac/source/src/server.cpp
@@ -931,7 +931,7 @@ void putflaginfo(ucharbuf &p, int flag)
             putint(p, f.actor_cn);
             break;
         case CTFF_DROPPED:
-            loopi(3) putuint(p, f.pos[i]*DMF);
+            loopi(3) putuint(p, (int)(f.pos[i]*DMF));
             break;
     }
 }
